Crate civicinfo2

Features

Handle the following Resources with ease from the central hub ...

divisions (search)

elections (election query and voter info query)

representatives (representative info by address and representative info by division)

Everything else about the Civic Info API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.elections().voter_info_query(...).doit()
let r = hub.elections().election_query(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
civicinfo2 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate civicinfo2;
use civicinfo2::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use civicinfo2::CivicInfo;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = CivicInfo::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.elections().voter_info_query("address")
             .official_only(false)
             .election_id("dolores")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

AdministrationRegion
Describes information about a regional election administrative area.

This type is not used in any activity, and only used as part of another schema.

AdministrativeBody
Information about an election administrative body (e.g. County Board of Elections).

This type is not used in any activity, and only used as part of another schema.

Candidate
Information about a candidate running for elected office.

This type is not used in any activity, and only used as part of another schema.

Channel
A social media or web channel for a candidate.

This type is not used in any activity, and only used as part of another schema.

CivicInfo
Central instance to access all CivicInfo related resource activities

Contest
Information about a contest that appears on a voter's ballot.

This type is not used in any activity, and only used as part of another schema.

DivisionMethodsBuilder
A builder providing access to all methods supported on division resources. It is not used directly, but through the CivicInfo hub.

DivisionSearchMethodBuilder
Searches for political divisions by their natural name or OCD ID.

DivisionSearchResponse
The result of a division search query.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

search divisions (response)

DivisionSearchResult
Represents a political geographic division that matches the requested query.

This type is not used in any activity, and only used as part of another schema.

Election
Information about the election that was queried.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

voter info query elections (none)

election query elections (none)

ElectionElectionQueryMethodBuilder
List of available elections to query.

ElectionMethodsBuilder
A builder providing access to all methods supported on election resources. It is not used directly, but through the CivicInfo hub.

ElectionOfficial
Information about individual election officials.

This type is not used in any activity, and only used as part of another schema.

ElectionVoterInfoQueryMethodBuilder
Looks up information relevant to a voter based on the voter's registered address.

ElectionsQueryResponse
The list of elections available for this version of the API.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

election query elections (response)

ElectoralDistrict
Describes the geographic scope of a contest.

This type is not used in any activity, and only used as part of another schema.

GeographicDivision
Describes a political geography.

This type is not used in any activity, and only used as part of another schema.

Office
Information about an Office held by one or more Officials.

This type is not used in any activity, and only used as part of another schema.

Official
Information about a person holding an elected office.

This type is not used in any activity, and only used as part of another schema.

PollingLocation
A location where a voter can vote. This may be an early vote site, an election day voting location, or a drop off location for a completed ballot.

This type is not used in any activity, and only used as part of another schema.

RepresentativeInfoData
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

representative info by division representatives (response)

RepresentativeInfoResponse
The result of a representative info lookup query.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

representative info by address representatives (response)

RepresentativeMethodsBuilder
A builder providing access to all methods supported on representative resources. It is not used directly, but through the CivicInfo hub.

RepresentativeRepresentativeInfoByAddresMethodBuilder
Looks up political geography and representative information for a single address.

RepresentativeRepresentativeInfoByDivisionMethodBuilder
Looks up representative information for a single geographic division.

SimpleAddressType
A simple representation of an address.

This type is not used in any activity, and only used as part of another schema.

Source
Contains information about the data source for the element containing it.

This type is not used in any activity, and only used as part of another schema.

VoterInfoResponse
The result of a voter info lookup query.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

voter info query elections (response)