Struct civicinfo2::ElectionVoterInfoQueryMethodBuilder

pub struct ElectionVoterInfoQueryMethodBuilder<'a, C, NC, A> where NC: 'a, C: 'a, A: 'a {
    // some fields omitted
}

Looks up information relevant to a voter based on the voter's registered address.

A builder for the voterInfoQuery method supported by a election resource. It is not used directly, but through a ElectionMethodsBuilder.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.elections().voter_info_query("address")
             .official_only(false)
             .election_id("labore")
             .doit();

Methods

impl<'a, C, NC, A> ElectionVoterInfoQueryMethodBuilder<'a, C, NC, A> where NC: NetworkConnector, C: BorrowMut<Client<NC>> + 'a, A: GetToken

fn doit(self) -> Result<VoterInfoResponse>

Perform the operation you have build so far.

fn address(self, new_value: &str) -> ElectionVoterInfoQueryMethodBuilder<'a, C, NC, A>

Sets the address query property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

The registered address of the voter to look up.

fn official_only(self, new_value: bool) -> ElectionVoterInfoQueryMethodBuilder<'a, C, NC, A>

Sets the official only query property to the given value.

If set to true, only data from official state sources will be returned.

fn election_id(self, new_value: &str) -> ElectionVoterInfoQueryMethodBuilder<'a, C, NC, A>

Sets the election id query property to the given value.

The unique ID of the election to look up. A list of election IDs can be obtained at https://www.googleapis.com/civicinfo/{version}/elections

fn delegate(self, new_value: &'a mut Delegate) -> ElectionVoterInfoQueryMethodBuilder<'a, C, NC, A>

Sets the delegate property to the given value.

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

fn param<T>(self, name: T, value: T) -> ElectionVoterInfoQueryMethodBuilder<'a, C, NC, A> where T: Str

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.